Refactor: RP-176 UserDTO 패키지 구조 변경
