+++ src/main/java/com/cherrypick/backend/domain/user/dto/request/UserRequestDTOs.java
@@ -1,0 +1,8 @@
+package com.cherrypick.backend.domain.user.dto.request;
+
+public class UserRequestDTOs
+{
+    public record DeleteRequestDTO(String reason){}
+
+
+}
