+++ src/main/java/com/cherrypick/backend/domain/user/dto/request/UserUpdateRequestDTO.java
@@ -1,11 +1,10 @@
-package com.cherrypick.backend.domain.user.dto;
+package com.cherrypick.backend.domain.user.dto.request;
 
 import com.cherrypick.backend.domain.user.enums.Gender;
 import com.cherrypick.backend.global.exception.BaseException;
 import com.cherrypick.backend.global.exception.enums.UserErrorCode;
 
 import java.time.LocalDate;
-import java.util.regex.Pattern;
 
 
 public record UserUpdateRequestDTO(
