+++ src/test/java/com/cherrypick/backend/domain/user/service/UserServiceTest.java
@@ -3,8 +3,8 @@ package com.cherrypick.backend.domain.user.service;
 import com.cherrypick.backend.domain.auth.domain.vo.AuthenticatedUser;
 import com.cherrypick.backend.domain.image.entity.Image;
 import com.cherrypick.backend.domain.image.service.ImageService;
-import com.cherrypick.backend.domain.user.dto.UserDetailResponseDTO;
-import com.cherrypick.backend.domain.user.dto.UserUpdateRequestDTO;
+import com.cherrypick.backend.domain.user.dto.response.UserDetailResponseDTO;
+import com.cherrypick.backend.domain.user.dto.request.UserUpdateRequestDTO;
 import com.cherrypick.backend.domain.user.entity.User;
 import com.cherrypick.backend.domain.user.enums.Gender;
 import com.cherrypick.backend.domain.user.enums.Role;
